Move IncomingRequestMiddleware
